[feature] Pixie - Block object edition per user type

+++ source/src/app/image-editor/canvas/canvas-keybinds.service.ts
@@ -2,6 +2,7 @@ import {Injectable} from '@angular/core';
 import {Keybinds} from 'common/core/keybinds/keybinds.service';
 import {ActiveObjectService} from './active-object/active-object.service';
 import {CanvasStateService} from './canvas-state.service';
+import { Settings } from '../../../common/core/config/settings.service';
 
 @Injectable()
 export class CanvasKeybindsService {
@@ -9,6 +10,7 @@ export class CanvasKeybindsService {
         private state: CanvasStateService,
         private keybinds: Keybinds,
         private activeObject: ActiveObjectService,
+        private config: Settings
     ) {}
 
     init() {
@@ -17,28 +19,28 @@ export class CanvasKeybindsService {
 
             this.keybinds.add('arrow_up', e => {
                 this.maybePreventDefault(e);
-                this.activeObject.move('top', -1);
+                if (this.config.get('pixie.isAdmin')) this.activeObject.move('top', -1);
             });
 
             this.keybinds.add('arrow_right', e => {
                 this.maybePreventDefault(e);
-                this.activeObject.move('left', 1);
+                if (this.config.get('pixie.isAdmin')) this.activeObject.move('left', 1);
             });
 
             this.keybinds.add('arrow_down', e => {
                 this.maybePreventDefault(e);
-                this.activeObject.move('top', 1);
+                if (this.config.get('pixie.isAdmin')) this.activeObject.move('top', 1);
             });
 
             this.keybinds.add('arrow_left', e => {
                 this.maybePreventDefault(e);
-                this.activeObject.move('left', -1);
+                if (this.config.get('pixie.isAdmin')) this.activeObject.move('left', -1);
             });
 
             this.keybinds.add('delete', e => {
                 this.maybePreventDefault(e);
                 if (this.activeObject.isEditing()) return;
-                this.activeObject.delete();
+                if (this.config.get('pixie.isAdmin')) this.activeObject.delete();
             });
         });
     }
